feat(devkit): add additionalShared helper
